refactor: nightly maintainability and docs
