+++ app/components/projects/ProjectsList.tsx
@@ -6,13 +6,11 @@ import { Plus } from "@/components/ui/icon";
 import { toast } from "sonner";
 import { createProject, deleteProject } from "@/lib/project/api";
 import type { ProjectRow } from "@/lib/project/api";
-import { ImageWithShimmer } from "@/lib/components/ImageWithShimmer";
 import UserProfile from "@/app/components/UserProfile";
 import { Button } from "@/components/ui/button";
 import { ConfirmDeleteDialog } from "@/components/ui/confirm-delete-dialog";
-import { DeleteButton } from "@/components/ui/delete-button";
 import { TooltipProvider } from "@/components/ui/tooltip";
-import { relativeTime } from "@/lib/project/relativeTime";
+import { ProjectCard } from "./ProjectCard";
 
 export default function ProjectsList({
 	initialProjects,
@@ -76,53 +74,11 @@ export default function ProjectsList({
 
 					<ul className="grid grid-cols-2 gap-x-4 gap-y-7 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6">
 						{projects.map((project) => (
-							<li key={project.id} className="group">
-								<div className="relative">
-									<button
-										type="button"
-										onClick={() => router.push(`/projects/${project.id}`)}
-										aria-label={`Open ${project.name}`}
-										className="block w-full overflow-hidden rounded-lg bg-muted focus-ring"
-									>
-										<div className="relative aspect-square">
-											{project.thumbnail_url ? (
-												<ImageWithShimmer
-													src={project.thumbnail_url}
-													alt={project.name}
-													fill
-													sizes="(min-width: 1280px) 200px, (min-width: 768px) 25vw, 50vw"
-													className="object-cover transition-transform duration-500 ease-out group-hover:scale-105"
-												/>
-											) : (
-												<div className="absolute inset-0 flex items-center justify-center text-label text-muted-foreground">
-													No preview
-												</div>
-											)}
-										</div>
-									</button>
-									<div className="absolute top-2 right-2 opacity-0 transition-opacity group-hover:opacity-100">
-										<DeleteButton
-											ariaLabel={`Delete ${project.name}`}
-											onClick={() => setDeleting(project)}
-										/>
-									</div>
-								</div>
-								<button
-									type="button"
-									onClick={() => router.push(`/projects/${project.id}`)}
-									className="mt-2 block w-full text-left"
-								>
-									<div className="truncate text-body font-semibold text-foreground">
-										{project.name}
-									</div>
-									<div
-										className="mt-0.5 truncate text-label text-muted-foreground"
-										suppressHydrationWarning
-									>
-										Edited {relativeTime(project.updated_at)}
-									</div>
-								</button>
-							</li>
+							<ProjectCard
+								key={project.id}
+								project={project}
+								onDelete={() => setDeleting(project)}
+							/>
 						))}
 					</ul>
 				</div>
